Fixes instance generation
There was a bug where after we switched to using What4 structs to represent
Lobot structs, the needed constraints on enumeration types (namely, that their
bitvector representation had a popcount of 1) were not getting propagated. This
fixes that by generating fresh constants for each of the fields, assuming those
constraints, and then building a struct out of those fresh constants.