Adds demo Lobot file to examples
This file (or something close to it) will be our target for the Piccolo build
demo. It demonstrates how we can use Lobot to organize the build of both the
Piccolo simulator and the RISC-V compliance suite with a common configuration,
and run all the tests through the simulator to verify that Piccolo is RISC-V
compliant.